// src/interfaces/PromoUp.interface.tsx
// Generated by https://quicktype.io

export interface PromoResponse {
  count: number;
  page: number;
  totalPages: number;
  data: PromoUp[];
}

export interface PromoUp {
  status: boolean;
  createdAt: string;
  updatedAt: string;
  image: Image;
  id: string;
}

export interface Image {
  url: string;
  id: string;
}