// src/interfaces/ErrorResp.interface.tsx
// Generated by https://quicktype.io

export interface ErrorResp {
  message: string;
  name: string;
  config: Config;
  code: string;
  status: null;
}

export interface Config {
  transitional: Transitional;
  transformRequest: null[];
  transformResponse: null[];
  timeout: number;
  xsrfCookieName: string;
  xsrfHeaderName: string;
  maxContentLength: number;
  maxBodyLength: number;
  env: Env;
  headers: Headers;
  baseURL: string;
  method: string;
  url: string;
  data: string;
}

export interface Env {
  FormData: null;
}

export interface Headers {
  Accept: string;
  'Content-Type': string;
}

export interface Transitional {
  silentJSONParsing: boolean;
  forcedJSONParsing: boolean;
  clarifyTimeoutError: boolean;
}