// src/interfaces/PromoDown.interface.tsx
import {Category} from './CategoryResponse.interface';
// Generated by https://quicktype.io

export interface PromoFinalResponse {
  count: number;
  page: number;
  totalPages: number;
  data: PromoFinal[];
}

export interface PromoFinal {
  status: boolean;
  createdAt: string;
  updatedAt: string;
  image: Image;
  category?: Category;
  id: string;
}

export interface Image {
  url: string;
  id: string;
}