// src/interfaces/CategoryResponse.interface.tsx
// Generated by https://quicktype.io

import {Image} from './Image.interface';
import {Node} from './Node.interface';
import {Subcategory} from './Subcategory.interface';

export interface CategoryResponse {
  count: number;
  page: number;
  totalPages: number;
  data: Category[];
}

export interface Category {
  subcategories: Partial<Subcategory>[];
  description: any[];
  status: boolean;
  cost: number;
  price: number;
  priceDiscount: number;
  currency: string;
  soldOut: boolean;
  recentProduct: string;
  name: string;
  subname: string;
  nodes: Partial<Node>[];
  createdAt: string;
  point: number;
  updatedAt: string;
  image: Image;
  id: string;
  ship: string;
  info: string[];
  aviableColors: string[];
}